[AArch64][PAC] Prevent immediate modifier substitution of AUT/PAC
Expand `blend(addr, imm)` in AsmPrinter, so that the discriminator
cannot be changed between blending `imm` in `addr` and using the
discriminator for signing or authenticating a pointer.

For that purpose, a new PAC pseudo instruction is introduced that has
separate operands for register and immediate modifiers. Instruction
selection is refactored along the way, so that AUT and PAC instructions
are selected by tablegen-erated patterns (always with a zero immediate
modifier at first) and a unified fixupBlendComponents function is used
by a custom inserter to fetch operands of `blend()` when possible. Aside
from unification, this helps inspecting operands of blend operation that
is located in another basic block which can be tricky for DAGISel.

This patch is somewhat related to
https://github.com/llvm/llvm-project/pull/132857.