[AArch64][PAC] Simplify pointer check emission (NFC)
The AArch64AsmPrinter::emitPtrauthCheckAuthenticatedValue method accepts
two operands, `bool ShouldTrap` and `const MCSymbol *OnFailure`, that
control the behavior of the emitted instruction sequence when the check
fails:
* `ShouldTrap` requests an error to be generated
* `OnFailure` requests branching to the given label after clearing the
  PAC field

An assertion in `emitPtrauthCheckAuthenticatedValue` ensures that
`OnFailure` must be null when `ShouldTrap` is set. But the opposite is
also true: when `ShouldTrap` is false, `OnFailure` is always non-null,
as otherwise the entire sequence would turn into a very expensive way
to ignore the result of the authentication ("check if the PAC field is
zero and make it zero if it is not yet").

This patch makes the invariant explicit by omitting the `ShouldTrap`
argument.